Login screen, User class, and Server
The Login screen displays the welcome, login, and create user screens while giving users the option to log in or create a new account. The server has access to a file that contains user info that allows users to log in under their accounts. The user class creates user objects from each user's information. The other files are for pictures and storing the user's information.